+++ client-sdks/ai-sdk/src/index.ts
@@ -2,5 +2,5 @@ export { chatRoute } from './chat-route';
 export type { chatRouteOptions } from './chat-route';
 
 export { toAISdkFormat } from './to-ai-sdk-format';
-export { WorkflowStreamToAISDKTransformer } from './to-ai-sdk-format';
-export type { WorkflowAiSDKType } from './to-ai-sdk-format';
+export { WorkflowStreamToAISDKTransformer } from './transformers';
+export type { WorkflowAiSDKType } from './transformers';
